### What's being changed:
smart sorting of MapPairs to avoid as many map-writes and reallocations as possible.

Note nDCG scores are identical to before.

before:
```
Min:    7.508042ms
Mean: 1.114266873s
Max:   5.62887875s
p50:  1.007076833s
p90:  2.439205375s
p99:  3.940842083s
```

after:
```
Min:    1.792375ms
Mean: 1.006939804s
Max:  3.240915833s
p50:  972.709041ms
p90:  2.391195417s
p99:  2.768915875s
```


### Review checklist

- [x] All new code is covered by tests where it is reasonable.
- [x] Performance tests have been run or not necessary.